+++ modules/work.nix
@@ -12,7 +12,7 @@
 
   hmModule = { pkgs, ... }: {
     home.packages = builtins.attrValues {
-      inherit (pkgs) awscli2 aws-vault mongodb-tools postman slack;
+      inherit (pkgs) awscli2 aws-vault mongodb-compass postman slack;
       inherit (pkgs.gnome) zenity;
     };
 
@@ -27,8 +27,9 @@
 
     programs.vscode.extensions = [
       pkgs.vscode-extensions.ethansk.restore-terminals
-      pkgs.vscode-extensions.rioj7.commandOnAllFiles
+      pkgs.vscode-extensions.ms-vsliveshare.vsliveshare
       pkgs.vscode-extensions.octref.vetur
+      pkgs.vscode-extensions.rioj7.commandOnAllFiles
     ];
 
     programs.vscode.userSettings = {