+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Added `vsliveshare` for `work`
+- Replaced `mongodb-tools` with `mongodb-compass` for `work`
 - Fixed `sigma` failing to boot by using an old version of `bcachefs`
 - Added `bcachefs` overlay to pin the version of `bcachefs`
 - Fixed delay when ending `neovim` macro recording in `VSCode`